requirements.txt
asgiref==3.8.1
Django==4.2.20
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
future==1.0.0
inflection==0.5.1
packaging==25.0
pillow==11.1.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzk==0.9
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
whitenoise==6.9.0
